+++ src/components/main-container/index.tsx
@@ -4,10 +4,11 @@ import { lazy, Suspense } from "react";
 import { Redirect, Route, Switch } from "wouter";
 const AboutPage = lazy(() => import("../about"));
 const CarsPage = lazy(() => import("../content/cars-page"));
-const TracksPage = lazy(() => import("../content/tracks-page"));
+const PrivacyPolicyPagePage = lazy(() => import("../privacy-policy"));
 const SeasonPage = lazy(() => import("../season/season-page"));
 const SeriesPage = lazy(() => import("../series/series-page"));
 const ShopPage = lazy(() => import("../shop-guide/shop-page"));
+const TracksPage = lazy(() => import("../content/tracks-page"));
 
 function MainContainer({ ...props }: StackProps) {
   usePageTracking();
@@ -41,6 +42,7 @@ function MainContainer({ ...props }: StackProps) {
             <Route path="/tracks" component={TracksPage} />
             <Route path="/checkout" component={ShopPage} />
             <Route path="/about" component={AboutPage} />
+            <Route path="/pp" component={PrivacyPolicyPagePage} />
 
             <Route>
               <Redirect to="/" />