+++ src/store/ui.ts
@@ -9,6 +9,7 @@ export enum ETabs {
   MyTracks = "/tracks",
   ShopGuide = "/checkout",
   About = "/about",
+  Privacy = "/pp",
 }
 
 export const useUiStorePersist = create(